review of the implementation and use of the macro case-word-size in the boot image code
The old definition of case-word-size from the library (vicare syntactic-extensions)
was making use of the value for "wordsize" form the library (vicare
installation-configuration), which is wrong when building 64-bit boot images
on a 32-bit system.

This fix reimplements case-word-size to use the value of wordsize from the
file "ikarus.config.ss".